+++ mods/BrewLAN_Plenae/NFTea/hook/lua/ui/dialogs/score.lua
@@ -1,0 +1,18 @@
+table.insert(tabs, {
+    title = "<LOC SCORE_crypto>Cryptocurrency",
+    button = "crypto",
+    scoreKey = 'crypto',
+    columns = {
+        {title = "<LOC SCORE_brewcoin>Brewcoin", scoreKey = 'Brewcoin',},
+        {title = "<LOC SCORE_teatherium>Teatherium", scoreKey = 'Teatherium',},
+        {title = "<LOC SCORE_sprouter>Sprouter", scoreKey = 'Sprouter',},
+        {title = "<LOC SCORE_dosh>Dosh", scoreKey = 'Dosh',},
+        {title = "<LOC SCORE_dosh>Dosh", scoreKey = 'Dosh',},
+        {title = "<LOC SCORE_dosh>Dosh", scoreKey = 'Dosh',},
+    },
+    types = {
+        {title = "<LOC SCORE_0027>Total", scoreKey = 'total',},
+        --{title = "<LOC SCORE_0028>Rate", scoreKey = 'rate',},
+    },
+    grid = true,
+})