+++ mods/BrewLAN_Plenae/NFTea/hook/lua/ui/help/tooltips.lua
@@ -1,0 +1,26 @@
+Tooltips['cycle_cryptocoin'] = {
+    title = "<LOC cycle_cryptocoin_title>Cycle coin",
+    description = "<LOC cycle_cryptocoin_desc>Swaps facility to mining a different coin.",
+    keyID = "cycle_cryptocoin",
+}
+
+Tooltips['PostScore_crypto'] = {
+    title = "<LOC PostScore_crypto_title>Cryptocurrency",
+    description = "<LOC PostScore_crypto_desc>Shows the cryptocurrency production totals per player",
+}
+Tooltips['PostScore_Teatherium'] = {
+    title = "<LOC PostScore_Teatherium_title>Teatherium",
+    description = "<LOC PostScore_Teatherium_desc>Shows the total Teatherium production per player",
+}
+Tooltips['PostScore_Brewcoin'] = {
+    title = "<LOC PostScore_Brewcoin_title>Brewcoin",
+    description = "<LOC PostScore_Brewcoin_desc>Shows the total Brewcoin production per player",
+}
+Tooltips['PostScore_Sprouter'] = {
+    title = "<LOC PostScore_Sprouter_title>Sprouter",
+    description = "<LOC PostScore_Sprouter_desc>Shows the total Sprouter production per player",
+}
+Tooltips['PostScore_Dosh'] = {
+    title = "<LOC PostScore_Dosh_title>Dosh",
+    description = "<LOC PostScore_Dosh_desc>Shows the total Dosh production per player",
+}