+++ mods/BrewLAN_Plenae/NFTea/hook/lua/ui/help/unitdescription.lua
@@ -1,0 +1,4 @@
+Description['sebn1ft'] = "<LOC cryptomint_build_desc>Converts energy into incredibly valuable cryptocurrency."
+Description['sabn1ft'] = "<LOC cryptomint_build_desc>Converts energy into incredibly valuable cryptocurrency."
+Description['srbn1ft'] = "<LOC cryptomint_build_desc>Converts energy into incredibly valuable cryptocurrency."
+Description['ssbn1ft'] = "<LOC cryptomint_build_desc>Converts energy into incredibly valuable cryptocurrency."