+++ mods/BrewLAN_Plenae/NFTea/units/SABN1FT/SABN1FT_script.lua
@@ -1,0 +1,7 @@
+local CryptoMintStructureUnit = import'/lua/defaultunits.lua'.CryptoMintStructureUnit
+
+SABN1FT = Class(CryptoMintStructureUnit) {
+    AmbientEffects = 'AT1PowerAmbient',
+}
+
+TypeClass = SABN1FT