+++ mods/BrewLAN_Plenae/NFTea/units/SEBN1FT/SEBN1FT_script.lua
@@ -1,0 +1,5 @@
+local CryptoMintStructureUnit = import'/lua/defaultunits.lua'.CryptoMintStructureUnit
+
+SEBN1FT = Class(CryptoMintStructureUnit) {}
+
+TypeClass = SEBN1FT