+++ mods/BrewLAN_Plenae/NFTea/units/SRBN1FT/SRBN1FT_script.lua
@@ -1,0 +1,5 @@
+local CryptoMintStructureUnit = import'/lua/defaultunits.lua'.CryptoMintStructureUnit
+
+SRBN1FT = Class(CryptoMintStructureUnit) {}
+
+TypeClass = SRBN1FT