+++ mods/BrewLAN_Plenae/NFTea/units/SSBN1FT/SSBN1FT_script.lua
@@ -1,0 +1,5 @@
+local CryptoMintStructureUnit = import'/lua/defaultunits.lua'.CryptoMintStructureUnit
+
+SSBN1FT = Class(CryptoMintStructureUnit) {}
+
+TypeClass = SSBN1FT