+++ mods/BrewLAN/documentation/Changelog.md
@@ -67,3 +67,7 @@ All changes by Sean Wheeldon (Balthazar) unless otherwise stated.
 ### 2022-03-30
 
 * Penetration: Finished the Cybran T3 Carpet Bomber.
+
+### 2022-03-31
+
+* NFTea: Stuff.